Using Eis random

+++ Dots2/src/Dots/Dot.cpp
@@ -2,7 +2,7 @@
 
 #include "../Env/Scene.h"
 
-Dot::Dot(const glm::vec2& pos, const uint8_t& id) : m_Pos(pos), m_Id(id)
+Dot::Dot(const glm::vec2& pos, const uint32_t& id) : m_Pos(pos), m_Id(id)
 {
 }
