+++ Dots2/src/Dots/Dot.h
@@ -8,7 +8,7 @@ class Dot
 {
 public:
 	Dot() = default;
-	Dot(const glm::vec2& pos, const uint8_t& id);
+	Dot(const glm::vec2& pos, const uint32_t& id);
 	~Dot() = default;
 
 	void Move(const glm::vec2& move, bool separateAxis = false);
@@ -16,12 +16,12 @@ public:
 
 	void SetPos(const glm::vec2& pos) { m_Pos = pos; }
 	inline glm::vec2 GetPos() const { return m_Pos; }
-	inline uint8_t GetId() const { return m_Id; }
+	inline uint32_t GetId() const { return m_Id; }
 
 private:
 	Brain brain;
 	glm::vec2 m_Pos;
-	uint8_t m_Id;
+	uint32_t m_Id;
 
 public:
 	bool drawn = false; // TEMP
